deleted unnecessary files and proper layouts made
