+++ resources/views/components/delete-modal.blade.php
@@ -1,0 +1,16 @@
+<div class="modal fade" id="deletConfirmation" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="deletConfirmationLabel" aria-hidden="true">
+    <div class="modal-dialog">
+    <div class="modal-content">
+        <div class="modal-header">
+        <h1 class="modal-title fs-5" id="deletConfirmationLabel">Are You Sure to Delete the {{ $module  }} ?? </h1>
+        <button type="button" class="btn-close btn-close-dlt" data-bs-dismiss="modal" aria-label="Close"></button>
+        </div>
+        <div class="modal-body">
+            {{ $message }}
+        </div>
+        <div class="modal-footer">
+        <button type="button" id="btn-dlt"  class="btn btn-dlt btn-danger rounded-2 px-3 py-0">Delete</button>
+        </div>
+    </div>
+    </div>
+</div>
