--- resources/views/components/input-label.blade.php
@@ -1,5 +1,0 @@
-@props(['value'])
-
-<label {{ $attributes->merge(['class' => 'block font-medium text-sm text-gray-700 dark:text-gray-300']) }}>
-    {{ $value ?? $slot }}
-</label>
