+++ resources/views/components/nav-bar.blade.php
@@ -10,7 +10,6 @@
         <a href="{{ route('logout') }}" onclick="event.preventDefault(); document.getElementById('logout-form').submit();"><i class="bi bi-box-arrow-right fs-5 "></i></a>
         <form id="logout-form" action="{{ route('employee') }}" method="GET" style="display: none;">
             @csrf
-
         </form>
     </div>
 </nav>
